Adding Maven Support
As well as fixing dependencies for JUnit 4 and 5.2 needed.
I needed to modify the settings of the project so we might lose support
for gradle.

I didn't know how to use it in the project as it was cloned :(

Anyways, I hope somebody can add the settings for using gradle again.